liblibc: don't use int/uint for intptr_t/uintptr_t
int/uint aren't considered FFI safe, replace them with the actual type they
represent (i64/u64 or i32/u32). This is a breaking change, but at most a cast
to `uint` or `int` needs to be added.

[breaking-change]

+++ src/liblibc/lib.rs
@@ -507,8 +507,8 @@ pub mod types {
             pub mod c99 {
                 pub type c_longlong = i64;
                 pub type c_ulonglong = u64;
-                pub type intptr_t = int;
-                pub type uintptr_t = uint;
+                pub type intptr_t = i32;
+                pub type uintptr_t = u32;
             }
             #[cfg(target_arch = "x86")]
             #[cfg(target_arch = "mips")]
@@ -702,8 +702,8 @@ pub mod types {
             pub mod c99 {
                 pub type c_longlong = i64;
                 pub type c_ulonglong = u64;
-                pub type intptr_t = int;
-                pub type uintptr_t = uint;
+                pub type intptr_t = i64;
+                pub type uintptr_t = u64;
             }
             pub mod posix88 {
                 pub type off_t = i64;
@@ -911,8 +911,8 @@ pub mod types {
             pub mod c99 {
                 pub type c_longlong = i64;
                 pub type c_ulonglong = u64;
-                pub type intptr_t = int;
-                pub type uintptr_t = uint;
+                pub type intptr_t = i64;
+                pub type uintptr_t = u64;
             }
             pub mod posix88 {
                 pub type off_t = i64;
@@ -1111,8 +1111,8 @@ pub mod types {
             pub mod c99 {
                 pub type c_longlong = i64;
                 pub type c_ulonglong = u64;
-                pub type intptr_t = int;
-                pub type uintptr_t = uint;
+                pub type intptr_t = i64;
+                pub type uintptr_t = u64;
             }
             pub mod posix88 {
                 pub type off_t = i64;
@@ -1229,7 +1229,7 @@ pub mod types {
             pub mod bsd44 {
                 use types::os::arch::c95::{c_char, c_int, c_uint, size_t, uintptr_t};
 
-                pub type SOCKET = uint;
+                pub type SOCKET = uintptr_t;
                 pub type socklen_t = c_int;
                 pub type sa_family_t = u16;
                 pub type in_port_t = u16;
@@ -1340,8 +1340,8 @@ pub mod types {
             pub mod c99 {
                 pub type c_longlong = i64;
                 pub type c_ulonglong = u64;
-                pub type intptr_t = int;
-                pub type uintptr_t = uint;
+                pub type intptr_t = i32;
+                pub type uintptr_t = u32;
             }
 
             pub mod posix88 {
@@ -1469,7 +1469,7 @@ pub mod types {
                     pub dwPageSize: DWORD,
                     pub lpMinimumApplicationAddress: LPVOID,
                     pub lpMaximumApplicationAddress: LPVOID,
-                    pub dwActiveProcessorMask: uint,
+                    pub dwActiveProcessorMask: uintptr_t,
                     pub dwNumberOfProcessors: DWORD,
                     pub dwProcessorType: DWORD,
                     pub dwAllocationGranularity: DWORD,
@@ -1703,8 +1703,8 @@ pub mod types {
             pub mod c99 {
                 pub type c_longlong = i64;
                 pub type c_ulonglong = u64;
-                pub type intptr_t = int;
-                pub type uintptr_t = uint;
+                pub type intptr_t = i32;
+                pub type uintptr_t = u32;
             }
             pub mod posix88 {
                 pub type off_t = i64;
@@ -1804,8 +1804,8 @@ pub mod types {
             pub mod c99 {
                 pub type c_longlong = i64;
                 pub type c_ulonglong = u64;
-                pub type intptr_t = int;
-                pub type uintptr_t = uint;
+                pub type intptr_t = i64;
+                pub type uintptr_t = u64;
             }
             pub mod posix88 {
                 pub type off_t = i64;
@@ -4382,7 +4382,7 @@ pub mod funcs {
                 pub fn glob(pattern: *const c_char,
                             flags: c_int,
                             errfunc: ::Nullable<extern "C" fn(epath: *const c_char,
-                                                              errno: int) -> int>,
+                                                              errno: c_int) -> int>,
                             pglob: *mut glob_t);
                 pub fn globfree(pglob: *mut glob_t);
             }
