+++ src/test/run-pass/extern-call-indirect.rs
@@ -22,17 +22,17 @@ mod rustrt {
 }
 
 extern fn cb(data: libc::uintptr_t) -> libc::uintptr_t {
-    if data == 1u {
+    if data == 1 {
         data
     } else {
-        fact(data - 1u) * data
+        fact(data - 1) * data
     }
 }
 
 fn fact(n: uint) -> uint {
     unsafe {
         println!("n = {}", n);
-        rustrt::rust_dbg_call(cb, n)
+        rustrt::rust_dbg_call(cb, n as libc::uintptr_t) as uint
     }
 }
